Move file upload logic to a separate store
